+++ example/zpoldbig_reduction.py
@@ -57,6 +57,8 @@ REG_INTEREST = None  # the region of interest of the array to plot in the format
 PLOT_FLUX = True  # flag whether to plot the flux of each beam
 PLOT_SHOW = False  # flag whether to show the figures, can slow down the reduction
 PLOT_SAVE = True  # flag whether to save the figures as png files
+ANALYZE = True  # flag whether to perform pixel performance analyze based on rms
+# and power spectrum
 
 # ======================= run the reduction pipeline ===========================
 
@@ -82,5 +84,6 @@ zpoldbig_result = z2pipl.reduce_zpoldbig(
         flat_flux=flat_flux, flat_err=flat_err, parallel=PARALLEL,
         return_pix_flag_list=True, table_save=TABLE_SAVE, plot=PLOT,
         plot_ts=PLOT_TS, reg_interest=REG_INTEREST, plot_flux=PLOT_FLUX,
-        plot_show=PLOT_SHOW, plot_save=PLOT_SAVE)
-zpoldbig_flux, zpoldbig_err, zpoldbig_pix_flag_list = zpoldbig_result[:2] + zpoldbig_result[-1:]
+        plot_show=PLOT_SHOW, plot_save=PLOT_SAVE, analyze=ANALYZE)
+zpoldbig_flux, zpoldbig_err, zpoldbig_pix_flag_list = \
+    zpoldbig_result[:2] + zpoldbig_result[-1:]